backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase==2.3.4
postgrest==0.14.0

# Database & Vector
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI & LLM Integration
openai==1.10.0  # For OpenRouter API compatibility
cohere==4.47.0
httpx==0.26.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.2.0
openpyxl==3.1.2
python-magic==0.4.27
Pillow==10.2.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development Tools
mypy==1.8.0
black==24.1.1
flake8==7.0.0
isort==5.13.2